% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.R
\name{capture_requests}
\alias{capture_requests}
\alias{start_capturing}
\alias{stop_capturing}
\title{Capture and record database transactions and save them as mocks}
\usage{
start_capturing(path)

stop_capturing()
}
\arguments{
\item{path}{the path to record mocks.}
}
\value{
NULL (invisibily)
}
\description{
Capture and record database transactions and save them as mocks
}
