% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture-requests.R
\name{check_db_path}
\alias{check_db_path}
\title{Check dittodb env path}
\usage{
check_db_path(.dittodb_env)
}
\arguments{
\item{.dittodb_env}{Environment object}
}
\value{
NULL, invisibly.
}
\description{
This is a function to help validate the correct steps are taken when running \code{\link[=start_db_capturing]{start_db_capturing()}}.
}
