% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{clean_statement}
\alias{clean_statement}
\title{Clean a statement string}
\usage{
clean_statement(string)
}
\arguments{
\item{string}{an SQL statement to clean}
}
\value{
the SQL statement stripped of extraneous bits
}
\description{
SQL statement strings sometimes have characters and specifications that don't
change the meaning or are determined at query time. To avoid this, before
hashing a statement we clean/strip these from the statement
}
\keyword{internal}
