% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_path_sanitize}
\alias{db_path_sanitize}
\title{Switch illegal characters for legal ones}
\usage{
db_path_sanitize(filename, replacement = "_")
}
\arguments{
\item{filename}{the file or folder to sanitize}

\item{replacement}{what should the illegal character(s) be replaced with?
(default: "_")}
}
\value{
the sanitized string
}
\description{
Inspired by the [fs](https://CRAN.R-project.org/package=fs) package's
`path_sanitize` function
}
\examples{
db_path_sanitize('this:string"has?issues')
}
\keyword{internal}
