% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbtest_debug_level}
\alias{dbtest_debug_level}
\title{Get the dbtest debug level and evaluate if it is above a level}
\usage{
dbtest_debug_level(level)
}
\arguments{
\item{level}{the level to test against (greater than or equal to)}
}
\value{
logical
}
\description{
Get the dbtest debug level and evaluate if it is above a level
}
\examples{
dbtest_debug_level(0)
dbtest_debug_level(2)
}
\keyword{internal}
