% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.R
\docType{data}
\name{.dbtest_env}
\alias{.dbtest_env}
\title{an environment for dbtest storing state}
\format{An object of class \code{environment} of length 0.}
\usage{
.dbtest_env
}
\description{
an environment for dbtest storing state
}
\keyword{datasets}
