% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittodb-env.R
\docType{data}
\name{.dittodb_env}
\alias{.dittodb_env}
\title{an environment for dittodb storing state}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.dittodb_env
}
\description{
an environment for dittodb storing state
}
\keyword{internal}
