% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-specific-connections.R
\name{driver-specifc-mock-connections}
\alias{driver-specifc-mock-connections}
\alias{DBIMockSQLiteConnection-class}
\alias{DBIMockRPostgreSQLConnection-class}
\alias{DBIMockRPostgresConnection-class}
\alias{DBIMockMariaDBConnection-class}
\title{Driver-specfic mock classes}
\description{
Each of the drivers that are supported have their own mock connection class.
They all inherit from DBIMockConnection as well as their own driver's connection
class. Each is only really available if the corresponding package is installed.
}
