% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture-requests.R
\name{extract_hash_from_object}
\alias{extract_hash_from_object}
\title{Extract a hash from a (result) object}
\usage{
extract_hash_from_object(obj)
}
\arguments{
\item{obj}{the DBI result object to hash}
}
\value{
hash for the object
}
\description{
This function should generally not be used, but must be exported for the
query recording function to work properly
}
\keyword{internal}
