% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.R
\name{get_redactor}
\alias{get_redactor}
\title{Get the current redactor}
\usage{
get_redactor()
}
\value{
the current list of columns to redact
}
\description{
This function should generally not be used, but must be exported for the
query recording function to work properly
}
\keyword{internal}
