% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_type}
\alias{get_type}
\title{Get the type of an SQL statement}
\usage{
get_type(statement)
}
\arguments{
\item{statement}{the statement to extract the first word from}
}
\value{
the first word in the statement
}
\description{
Get the type of an SQL statement
}
\keyword{internal}
