% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{hash}
\alias{hash}
\title{Make a (short) hash from a string}
\usage{
hash(string, n = 6)
}
\arguments{
\item{string}{the string to hash}

\item{n}{how long should the hash be? (default: 6)}
}
\value{
a hash for the string
}
\description{
Make a (short) hash from a string
}
\keyword{internal}
