% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{make_path}
\alias{make_path}
\title{make a mock path}
\usage{
make_path(path, type, hash)
}
\arguments{
\item{path}{the path to look in}

\item{type}{what type of query is it? (e.g. \code{SELECT}, \code{INSERT})}

\item{hash}{the hash of the query}
}
\value{
a constructed path to a mock
}
\description{
make a mock path
}
\keyword{internal}
