% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBconnection.R
\docType{methods}
\name{mock-db-methods}
\alias{mock-db-methods}
\alias{dbSendQuery,DBIMockConnection,character-method}
\alias{dbFetch,DBIMockResult-method}
\alias{dbClearResult,DBIMockResult-method}
\alias{dbDisconnect,DBIMockConnection-method}
\title{Methods for interacting with DB mocks instead of an actual database}
\usage{
\S4method{dbSendQuery}{DBIMockConnection,character}(conn, statement, ...)

\S4method{dbFetch}{DBIMockResult}(res, n = -1, ...)

\S4method{dbClearResult}{DBIMockResult}(res, n, ...)

\S4method{dbDisconnect}{DBIMockConnection}(conn, ...)
}
\arguments{
\item{conn}{a database connection (for dispatch with these methods, it should
be of class `DBIMockConnection`)}

\item{statement}{an SQL statement to execute}

\item{...}{arguments passed on inside of the methods}

\item{res}{a result object (for dispatch with these methods, it should
be of class `DBIMockResult`)}

\item{n}{number of results to fetch (ignored)}
}
\description{
Various methods (`dbSendQuery`, `dbFetchQuery`) that are mocks of the DBI
methods of the same name. Instead of actually interacting with a database,
they read in mock responses and the code proceeds after that.
}
