% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycflights13-sql.R
\name{nycflights13_create_sql}
\alias{nycflights13_create_sql}
\title{Create a standardised database for testing}
\usage{
nycflights13_create_sql(con, schema = "", ...)
}
\arguments{
\item{con}{an SQL connection (i.e a PostgreSQL connection)}

\item{schema}{schema to write the tables ("", or no schema by default)}

\item{...}{additional parameters to connect to a database}
}
\value{
the connection given in \code{con} invisibly, generally called for the
side effects of writing to the database
}
\description{
Using the connection given in \code{con}, create a database including a few tables
from the \url{nycflights13} dataset.
}
\examples{
\dontrun{
connections <- list(
 odbc = DBI::dbConnect(
  odbc::odbc(),
  Driver   = "PostgreSQL Unicode",
  Server   = "127.0.0.1",
  Database = "postgres",
  UID      = "travis",
  PWD      = "",
  Port     = 5432
 ),
 rpostgresql = RPostgreSQL::dbConnect(
  drv      = DBI::dbDriver("PostgreSQL"),
  host     = "127.0.0.1",
  dbname   = "postgres",
  user     = "travis",
  password = "",
  port     = 5432
 ),
 rpostgres = DBI::dbConnect(
  drv      = RPostgres::Postgres(),
  host     = "127.0.0.1",
  dbname   = "postgres",
  user     = "travis",
  password = "",
  port     = 5432
 )
)

nycflights13_create_sql(connections$odbc, schema = "nycflights13")
# same as
# nycflights13_create_sql(connections$rpostgresql, schema = "nycflights13")
# also same as
# nycflights13_create_sql(connections$rpostgres, schema = "nycflights13")
}
}
