% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycflights13_sql.R
\name{nycflights13_sql}
\alias{nycflights13_sql}
\title{Create a standardised database for testing}
\usage{
nycflights13_sql(con, schema = "", method = c("dbi", "dplyr"), ...)
}
\arguments{
\item{con}{an SQL connection (i.e a PostgreSQL connection)}

\item{schema}{schema to write the tables ("", or no schema by default)}

\item{method}{should the tables be created with DBI methods or dplyr methods?
(either "dbi" or "dplyr"; "dbi" by default)}

\item{...}{additional parameters to connect to a database}
}
\value{
the connection given in `con` invisibly, generally called for the side effects
of writing to the  database
}
\description{
Using the connection given in `con`, create a database including a few tables
from the [`nycflights2013`] dataset.
}
\examples{
\dontrun{
# psql_con is a PostgreSQL connnection from .Rprofile
nycflights13_sql(psql_con, schema = "nycflights13")
}
}
