% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{path_sanitize}
\alias{path_sanitize}
\alias{db_path_sanitize}
\title{Sanitize a filename by removing directory paths and invalid characters (from the `fs` package)}
\usage{
path_sanitize(filename, replacement = "")

db_path_sanitize(filename, replacement = "_")
}
\arguments{
\item{filename}{A character vector to be sanitized.}

\item{replacement}{A character vector used to replace invalid characters.}
}
\description{
`path_sanitize()` removes the following:
- [Control characters](https://en.wikipedia.org/wiki/C0_and_C1_control_codes)
- [Reserved characters](https://kb.acronis.com/content/39790)
- Unix reserved filenames (`.` and `..`)
- Trailing periods and spaces (invalid on Windows)
- Windows reserved filenames (`CON`, `PRN`, `AUX`, `NUL`, `COM1`, `COM2`,
  `COM3`, COM4, `COM5`, `COM6`, `COM7`, `COM8`, `COM9`, `LPT1`, `LPT2`,
  `LPT3`, `LPT4`, `LPT5`, `LPT6`, LPT7, `LPT8`, and `LPT9`)
The resulting string is then truncated to [255 bytes in length](https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits)
}
\seealso{
<https://CRAN.R-project.org/package=fs>, where this function was
taken from.
}
