% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact.R
\name{redact_columns}
\alias{redact_columns}
\title{Redact columns from a dataframe with the default redactors}
\usage{
redact_columns(data, columns, ignore.case = TRUE, ...)
}
\arguments{
\item{data}{a dataframe to redact}

\item{columns}{character, the columns to redact}

\item{ignore.case}{should case be ignored? (default: \code{TRUE})}

\item{...}{additional options to pass on to \code{grep}}
}
\value{
data, with the columns specified in \code{columns} duly redacted
}
\description{
Redact columns from a dataframe with the default redactors
}
