% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_table_id}
\alias{sanitize_table_id}
\title{Sanitize Table id}
\usage{
sanitize_table_id(id, ...)
}
\arguments{
\item{id}{the table identifier (an \code{Id}, a vector of strings, or a string)}

\item{...}{additional arguments (to allow for things like \code{schema_name} that
\code{odbc} uses.)}
}
\value{
the first word in the statement
}
\description{
Tables are identified and specified with a large number of ways across
drivers. For the purposes of {dittodb}, the details are less important since
we almost always just want a flat representation (\emph{ie} for filenames). This
takes the various formats and returns a string with various elements
separated by dots.
}
\keyword{internal}
