% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize-bit64.R
\name{serialize_bit64}
\alias{serialize_bit64}
\title{Make a data.frame with pre-serialized integer64 columns}
\usage{
serialize_bit64(df_in)
}
\arguments{
\item{df_in}{the data.frame to pre-serialize \code{integer64} columns from}
}
\value{
data.frame where each \code{integer64} column is a serialized expression of itself
}
\description{
\code{integer64} types do not \code{dput} wells with \code{hexNumeric}. Instead, we serialize and
then unserialize them cf truecluster/bit64#27 and store that as one expression per column
This makes the resulting data.frame awkward, but it will be recorded by \code{dput} in a
way that does get read back in faithfully (so long as the expression isn't quoted).
}
\keyword{internal}
