% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-db.R
\name{with_mock_dbi}
\alias{with_mock_dbi}
\title{Run the DBI queries inside against a mocked db}
\usage{
with_mock_dbi(expr)
}
\arguments{
\item{expr}{the expression to execute}
}
\value{
nothing
}
\description{
Run the DBI queries inside against a mocked db
}
