% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture-requests.R
\name{check_db_path}
\alias{check_db_path}
\title{Check dittodb environment path}
\usage{
check_db_path(.dittodb_env, n = 1L)
}
\arguments{
\item{.dittodb_env}{Environment object}

\item{n}{Integer. Type of error to throw.}
}
\value{
NULL, invisibly.
}
\description{
This function should generally not be used, but must be exported for the
query recording function to work properly
}
