% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/result.R, R/listTables.R
\docType{class}
\name{mock-db-methods}
\alias{mock-db-methods}
\alias{DBIMockConnection-class}
\alias{dbDisconnect,DBIMockConnection-method}
\alias{dbListTables,DBIMockConnection-method}
\alias{DBIMockResult-class}
\alias{dbSendQuery,DBIMockConnection,character-method}
\alias{dbFetch,DBIMockResult-method}
\alias{fetch,DBIMockResult-method}
\alias{dbClearResult,DBIMockResult-method}
\alias{dbHasCompleted,DBIMockResult-method}
\alias{dbListTables,DBIMockODBCConnection-method}
\title{Methods for interacting with DB mocks instead of an actual database}
\usage{
\S4method{dbDisconnect}{DBIMockConnection}(conn, ...)

\S4method{dbListTables}{DBIMockConnection}(conn, ...)

\S4method{dbSendQuery}{DBIMockConnection,character}(conn, statement, ...)

\S4method{dbFetch}{DBIMockResult}(res, n = -1, ...)

\S4method{fetch}{DBIMockResult}(res, n = -1, ...)

\S4method{dbClearResult}{DBIMockResult}(res, n, ...)

\S4method{dbHasCompleted}{DBIMockResult}(res, ...)

\S4method{dbListTables}{DBIMockODBCConnection}(conn, ...)
}
\arguments{
\item{conn}{a database connection (for dispatch with these methods, it should
be of class \code{DBIMockConnection})}

\item{...}{arguments passed on inside of the methods}

\item{statement}{an SQL statement to execute}

\item{res}{a result object (for dispatch with these methods, it should
be of class \code{DBIMockResult})}

\item{n}{number of results to fetch (ignored)}
}
\description{
Various methods (\code{dbSendQuery}, \code{dbFetchQuery}) that are mocks of the DBI
methods of the same name. Instead of actually interacting with a database,
they read in mock responses and the code proceeds after that.
}
