% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/listTables.R, R/quote.R,
%   R/result.R
\docType{class}
\name{mock-db-methods}
\alias{mock-db-methods}
\alias{DBIMockConnection-class}
\alias{dbDisconnect,DBIMockConnection-method}
\alias{dbListTables,DBIMockConnection-method}
\alias{dbListFields,DBIMockConnection,character-method}
\alias{dbListFields,DBIMockConnection,Id-method}
\alias{dbListFields,DBIMockConnection,ANY-method}
\alias{dbQuoteIdentifier,DBIMockConnection,character-method}
\alias{dbQuoteIdentifier,DBIMockConnection,SQL-method}
\alias{DBIMockResult-class}
\alias{dbSendQuery,DBIMockConnection,character-method}
\alias{dbFetch,DBIMockResult-method}
\alias{fetch,DBIMockResult,ANY-method}
\alias{fetch,DBIMockResult-method}
\alias{dbClearResult,DBIMockResult-method}
\alias{dbHasCompleted,DBIMockResult-method}
\alias{dbGetQuery,DBIMockConnection,character-method}
\title{Methods for interacting with DB mocks instead of an actual database}
\usage{
\S4method{dbDisconnect}{DBIMockConnection}(conn, ...)

\S4method{dbListTables}{DBIMockConnection}(conn, ...)

\S4method{dbListTables}{DBIMockConnection}(conn, ...)

\S4method{dbListFields}{DBIMockConnection,character}(conn, name, ...)

\S4method{dbListFields}{DBIMockConnection,Id}(conn, name, ...)

\S4method{dbListFields}{DBIMockConnection,ANY}(conn, name, ...)

\S4method{dbQuoteIdentifier}{DBIMockConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{DBIMockConnection,SQL}(conn, x, ...)

\S4method{dbSendQuery}{DBIMockConnection,character}(conn, statement, ...)

\S4method{dbFetch}{DBIMockResult}(res, n = -1, ...)

\S4method{fetch}{DBIMockResult,ANY}(res, n = -1, ...)

\S4method{dbClearResult}{DBIMockResult}(res, n, ...)

\S4method{dbHasCompleted}{DBIMockResult}(res, ...)

\S4method{dbGetQuery}{DBIMockConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{a database connection (for dispatch with these methods, it should
be of class \code{DBIMockConnection})}

\item{...}{arguments passed on inside of the methods}

\item{name}{name of the table (for \code{\link{dbListFields}})}

\item{x}{a name to quote (for \code{\link{dbQuoteIdentifier}})}

\item{statement}{an SQL statement to execute}

\item{res}{a result object (for dispatch with these methods, it should
be of class \code{DBIMockResult})}

\item{n}{number of results to fetch (ignored)}
}
\description{
Various methods (\code{dbSendQuery}, \code{dbFetchQuery}) that are mocks of the DBI
methods of the same name. Instead of actually interacting with a database,
they read in mock responses and the code proceeds after that.
}
