% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycflights13_sqlite.R
\name{nycflights13_sql}
\alias{nycflights13_sql}
\title{Create an in-memory SQLite database for testing}
\usage{
nycflights13_sql(sqlite = TRUE, use = "DBI", ...)
}
\arguments{
\item{sqlite}{logical, when FALSE the function shall use odbc}

\item{use}{should the connection use/return a dplyr-based or DBI-based
connection? (default: DBI)}

\item{...}{additional parameters to connect to a database}
}
\value{
RSQLite object
}
\description{
Create an in-memory SQLite database for testing
}
\examples{
\dontrun{
nycflights13_sql(type = "sqlite")
}
}
