% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycflights13_sql.R
\name{nycflights13_sqlite}
\alias{nycflights13_sqlite}
\title{Create an in-memory SQLite database for testing}
\usage{
nycflights13_sqlite(...)
}
\arguments{
\item{...}{additional parameters to connect to a database (most are passed on
to [`nycflights_sql`])}
}
\value{
RSQLiteConnection
}
\description{
Create an in-memory SQLite database for testing
}
\examples{
\dontrun{
con <- nycflights13_sqlite()
}
}
