% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycflights13_sqlite.R
\name{nycflights13_sqlite}
\alias{nycflights13_sqlite}
\title{Create an in-memory SQLite database for testing}
\usage{
nycflights13_sqlite()
}
\value{
A xxx object
}
\description{
Create an in-memory SQLite database for testing
}
\examples{
nycflights13_sqlite()
}
