% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact.R
\name{redact_columns}
\alias{redact_columns}
\title{Redact columns from a dataframe with the default redactors}
\usage{
redact_columns(data, columns, ignore.case = TRUE, ...)
}
\arguments{
\item{data}{a dataframe to redact}

\item{columns}{character, the columns to redact}

\item{ignore.case}{should case be ignored? (default: \code{TRUE})}

\item{...}{additional options to pass on to \code{grep} when matching the column
names}
}
\value{
data, with the columns specified in \code{columns} duly redacted
}
\description{
This function redacts the columns specified in \code{columns} in the data given in
\code{data} using dbtest's standard redactors.
}
\details{
The standard redactors replace all values in the column with the following
values based on the columns type:
\itemize{
\item integer -- \code{9L}
\item numeric -- \code{9}
\item character -- \code{"[redacted]"}
\item \code{POSIXct} (date times) -- \code{as.POSIXct("1988-10-11T17:00:00", tz = tzone)}
}
}
