% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-paths.R
\name{with_mock_path}
\alias{with_mock_path}
\title{Run the DBI queries in an alternate mock directory}
\usage{
with_mock_path(path, expr, replace = FALSE)
}
\arguments{
\item{path}{the alternate directory}

\item{expr}{the expression to execute}

\item{replace}{foobar}
}
\value{
nothing
}
\description{
When testing with dittodb, wrap your tests in \code{with_mock_path({})} to use the
database fixtures located in other directories. \code{dittodb} will look for
fixtures in the directory specified by the user, which can be a temporal
or permanent location.
}
